Make tests on type checking unit tests
