+++ parselglossy/types.py
@@ -26,88 +26,88 @@
 # parselglossy library, see: <http://parselglossy.readthedocs.io/>
 #
 
-from typing import List, Optional, Tuple
+import re
+from typing import Any, Callable, List, Optional, Tuple, Union
 
-from .exceptions import Error, ParselglossyError, collate_errors
-from .utils import JSONDict, type_fixers, type_matches
+RunCallable = Callable[[Any, str], Tuple[str, Any]]
+
+ScalarTypes = Union[bool, str, int, float, complex]
+
+allowed_scalar_types = ["str", "int", "float", "complex", "bool"]
 
 
-def typenade(incoming: JSONDict, types: JSONDict) -> Optional[JSONDict]:
-    """Checks types of input values for a merge input ``dict``.
+def _type_check_scalar(value: ScalarTypes, expected_type: str) -> bool:
+    return type(value).__name__ == expected_type
+
+
+ListTypes = Union[List[bool], List[str], List[int], List[float], List[complex]]
+
+allowed_list_types = ["List[{}]".format(t) for t in allowed_scalar_types]
+
+
+def _type_check_list(value: ListTypes, expected_type: str) -> bool:
+    # make sure that value is actually a list
+    if type(value).__name__ == "list":
+        # iterate over each element of the list
+        # and check whether it matches T
+        type_checks = all((_type_check_scalar(x, expected_type) for x in value))
+    else:
+        type_checks = False
+
+    return type_checks
+
+
+AllowedTypes = Union[ScalarTypes, ListTypes]
+
+allowed_types = allowed_scalar_types + allowed_list_types
+
+
+def type_matches(value: AllowedTypes, expected_type: str) -> Optional[bool]:
+    """Checks whether a value is of the expected type.
 
     Parameters
     ----------
-    incoming: JSONDict
-        The input `dict`. This is supposed to be the one obtained by merging
-        user and template `dict`-s.
-    types: JSONDict
-        Types of all keywords in the input. Generated from :func:`view_by_types`.
-
-    Returns
-    -------
-    outgoing: JSONDict
-        A dictionary with all values type checked.
-
-    Raises
-    ------
-    :exc:`ValidationError`
+    value : AllowedTypes
+      Value whose type needs to be checked
+    expected_type : str
 
     Notes
     -----
-    This is porcelain over the recursive function :func:`rec_typenade`.
-    """
-    outgoing, errors = rec_typenade(incoming, types)
-
-    if errors:
-        msg = collate_errors(when="checking types", errors=errors)
-        raise ParselglossyError(msg)
-
-    return outgoing
-
-
-def rec_typenade(
-    incoming: JSONDict, types: JSONDict, *, address: Tuple = ()
-) -> Tuple[JSONDict, List[Error]]:
-    """Perform type checking and type fixing.
-
-    Parameters
-    ----------
-    incoming: JSONDict
-        The input `dict`. This is supposed to be the one obtained by merging
-        user and template `dict`-s.
-    types: JSONDict
-        Types of all keywords in the input. Generated from :func:`view_by_types`.
-    address: Tuple[str]
-        A tuple of keys need to index the current value in the recursion. See
-        Notes.
+    Allowed types T are: `str`, `int`, `float`, `complex`, `bool`,
+    as well as `List[T]`.
 
     Returns
     -------
-    outgoing: JSONDict
-        A dictionary with all default values fixed.
-    errors_at: List[Tuple[str]]
-        A list of keys to access elements in the `dict` that raised an error.
-        See Notes.
+    True if value has the type expected_type, otherwise False.
+
+    Raises
+    ------
+    ValueError
+        If expected_type is not among the allowed types.
     """
 
-    outgoing = {}
-    errors = []
+    # first verify whether expected_type is allowed
+    if expected_type not in allowed_types:
+        raise ValueError("could not recognize expected_type: {}".format(expected_type))
 
-    for k, v in incoming.items():
-        if not isinstance(v, dict):
-            t = types[k]
-            if type_matches(incoming[k], t):
-                outgoing[k] = type_fixers[t](incoming[k])
-            else:
-                msg = "Actual ({0}) and declared ({1}) types do not match".format(
-                    type(incoming[k]).__name__, t
-                )
-                errors.append(Error(address + (k,), msg))
-                outgoing[k] = None  # type: ignore
-        else:
-            outgoing[k], errs = rec_typenade(
-                incoming=v, types=types[k], address=(address + (k,))
-            )
-            errors.extend(errs)
+    expected_type_is_list = re.search(r"^List\[(\w+)\]$", expected_type)
 
-    return outgoing, errors
+    if expected_type_is_list is not None:
+        return _type_check_list(value, expected_type_is_list.group(1))  # type: ignore
+    else:
+        return _type_check_scalar(value, expected_type)  # type: ignore
+
+
+type_fixers = {
+    "bool": bool,
+    "complex": complex,
+    "float": float,
+    "int": int,
+    "str": str,
+    "List[bool]": lambda x: list(map(bool, x)),
+    "List[complex]": lambda x: list(map(complex, x)),
+    "List[float]": lambda x: list(map(float, x)),
+    "List[int]": lambda x: list(map(int, x)),
+    "List[str]": lambda x: list(map(str, x)),
+}  # type: Dict[str, Callable[[Any], Any]]
+"""Dict[str, Callable[[Any], Any]]: dictionary holding functions for type fixation."""
