+++ setup.cfg
@@ -15,7 +15,7 @@ replace = __version__ = '{new_version}'
 universal = 1
 
 [flake8]
-exclude = docs
+exclude = docs,tests
 max-line-length = 88
 
 [pycodestyle]
